+++ src/common/types/plugin.ts
@@ -14,7 +14,8 @@ export interface PluginPackage {
     },
     config?: {
         system?: {
-            permissions?: string[]
+            permissions?: string[],
+            deprecated?: boolean;
         }
     }
 }