+++ src/renderer/load-replay.ts
@@ -37,6 +37,7 @@ import { sanityCheckCommands, writeCommands } from "./process-replay/write-comma
 import getContainerSize from "./process-replay/get-container-size";
 import { setDumpUnitCall } from "./plugins/plugin-system-ui";
 import { calculateImagesFromSpritesIscript } from "./iscript/images-from-iscript";
+import { CMDS } from "./process-replay/commands/commands";
 
 export default async (filepath: string) => {
   gameStore().disposeGame();
@@ -132,7 +133,7 @@ export default async (filepath: string) => {
 
   const terrain = await chkToTerrainMesh(
     map, {
-      //TODO: replace since HD2 and HD will be loaded
+    //TODO: replace since HD2 and HD will be loaded
     textureResolution: settings.assets.terrain === AssetTextureResolution.SD ? UnitTileScale.SD : UnitTileScale.HD,
     anisotropy: settings.graphics.anisotropy
   }
@@ -184,9 +185,23 @@ export default async (filepath: string) => {
   processStore().increment(Process.ReplayInitialization);
   ImageHD.useDepth = false;
 
-  const unitSprites = map.units.map(u => u.sprite).filter(s => Number.isInteger(s)) as number[];
-  const mapSprites = calculateImagesFromSpritesIscript(assets.bwDat, [...unitSprites, ...map.sprites.map(s => s.spriteId)]);
-  await Promise.all(mapSprites.map((spriteId) => assets.loadAnim(spriteId, settings.assets.images === AssetTextureResolution.SD ? UnitTileScale.SD : UnitTileScale.HD2)));
+  const preloadCommands = new CommandsStream(replay.rawCmds, replay.stormPlayerToGamePlayer);
+  const preloadCommandTypes = [CMDS.TRAIN.id, CMDS.UNIT_MORPH.id, CMDS.BUILDING_MORPH.id, CMDS.BUILD.id];
+  const preloadCommandUnits = new Set<number>();
+
+  for (const command of preloadCommands.generate()) {
+    if (typeof command !== "number") {
+      if (preloadCommandTypes.includes(command.id)) {
+        preloadCommandUnits.add(command.unitTypeId!);
+      }
+    }
+  }
+
+  const unitSprites = new Set(map.units.map(u => u.sprite).filter(s => Number.isInteger(s)) as number[]);
+  const allSprites = [...preloadCommandUnits, ...unitSprites, ...new Set(map.sprites.map(s => s.spriteId))];
+  const allImages = calculateImagesFromSpritesIscript(assets.bwDat, allSprites);
+
+  await Promise.all(allImages.map((imageId) => assets.loadAnim(imageId, settings.assets.images === AssetTextureResolution.SD ? UnitTileScale.SD : UnitTileScale.HD2)));
 
   const disposeGame = await TitanReactorGame(
     map,